integrations/requirements.txt
# Core dependencies
celery>=5.3.0
redis>=4.5.0
supabase>=1.0.0

# AI/ML Libraries
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0
huggingface-hub>=0.30.0,<1.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.8.0
numpy>=1.24.0

# Video Processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.8.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0

# Environment and Config
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging and Monitoring
loguru>=0.7.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 